split up large etl files into smaller resuable chunks
