+++ web/sites/ww2visualiser.co.uk/src/etl/lib/constants.ts
@@ -1,0 +1,6 @@
+export const SPARQL_ENDPOINT = "https://query.wikidata.org/sparql";
+export const WP_API = "https://en.wikipedia.org/w/api.php";
+export const WP_SUMMARY_API = "https://en.wikipedia.org/api/rest_v1/page/summary";
+
+export const UA =
+  "WW2Visualiser/1.0 (build-time data fetcher; contact james_d02@protonmail.com)";
